# openapi.yaml
openapi: 3.1.0
info:
  title: Community Platform API
  description: |
    Complete REST API for a Next.js community platform with social features, vendor marketplace, 
    bookings, gamification, and admin capabilities.
    
    ## Features
    - 🔐 Authentication via Clerk
    - 🛡️ Rate limiting with Upstash Redis
    - ✅ Zod validation on all inputs
    - 📊 Supabase PostgreSQL with RLS
    - 💳 Stripe payment integration
    - 📧 Webhook support
    
    ## Authentication
    All authenticated endpoints require a valid Clerk session token in the Authorization header:
    ```
    Authorization: Bearer <clerk_jwt_token>
    ```
    
    ## Rate Limiting
    Rate limits vary by endpoint and authentication status:
    - Anonymous read: 20 requests/minute
    - Anonymous write: 5 requests/minute
    - Authenticated read: 100 requests/minute
    - Authenticated write: 30 requests/minute
    - Admin operations: 10 requests/minute
    - Webhooks: 100 requests/minute
    
    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests in window
    - `X-RateLimit-Reset`: Timestamp when limit resets
    
    ## Error Responses
    All errors follow a consistent format:
    ```json
    {
      "error": "Error message",
      "details": "Additional error details or validation errors"
    }
    ```
    
    Common HTTP status codes:
    - `400` Bad Request - Invalid input or validation error
...
    
...
servers:
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Clerk JWT token obtained from user authentication.
        Include in Authorization header: `Authorization: Bearer <token>`
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Validation failed
        details:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
          description: Additional error details or validation errors
          example:
            - field: content
              message: String must contain at least 1 character(s)
    ValidationError:
      type: object
      required:
        - error
        - details
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
                description: Field that failed validation
                example: content
              message:
                type: string
                description: Validation error message
                example: String must contain at least 1 character(s)
    RateLimitError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Rate limit exceeded
        retryAfter:
          type: integer
          description: Seconds until rate limit resets
          example: 45
    Comment:
      type: object
      required:
        - id
        - post_id
        - user_id
        - content
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique comment identifier
        post_id:
          type: string
          format: uuid
          description: ID of the post this comment belongs to
        user_id:
          type: string
          description: Clerk user ID of comment author
        content:
          type: string
          minLength: 1
          maxLength: 1000
          description: Comment text content
        parent_comment_id:
          type: string
          format: uuid
          nullable: true
          description: ID of parent comment for nested replies
        is_deleted:
          type: boolean
          default: false
          description: Soft delete flag
        created_at:
          type: string
          format: date-time
          description: Comment creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        author:
          $ref: '#/components/schemas/ProfileSummary'
        likes:
          type: array
          items:
            type: object
            properties:
              count:
                type: integer
                description: Total number of likes
    CommentCreate:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
          description: Comment text content
          example: Great post! Very informative.
        parent_id:
          type: string
          format: uuid
          nullable: true
          description: Parent comment ID for nested replies
          example: null
    ProfileSummary:
      type: object
      required:
        - id
        - username
      properties:
        id:
          type: string
          format: uuid
          description: Profile ID
        username:
          type: string
          description: User's username
        display_name:
          type: string
          description: User's display name
        avatar_url:
          type: string
          format: uri
          nullable: true
          description: Profile picture URL
        is_verified:
          type: boolean
          description: User verification status
        vendor_verified:
          type: boolean
          description: Vendor verification status
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized. Please sign in.
...
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Resource not found
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Maximum requests allowed
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
  parameters:
    PostId:
      name: id
      in: path
      required: true
      description: Post UUID
      schema:
        type: string
        format: uuid
      example: 123e4567-e89b-12d3-a456-426614174000
...
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Offset:
      name: offset
      in: query
      description: Number of items to skip (for pagination)
      schema:
        type: integer
        minimum: 0
        default: 0
paths:
  /api/posts/{id}/comments:
    get:
      tags:
        - Comments
      summary: Get all comments for a post
      description: |
        Retrieve paginated comments for a specific post. Includes author information and like counts.
        Supports nested comments via parent_comment_id.
      operationId: getPostComments
      parameters:
        - $ref: '#/components/parameters/PostId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Comments retrieved successfully
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Rate limit (20 anonymous, 100 authenticated)
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                required:
                  - comments
                  - total
                  - limit
                  - offset
                properties:
                  comments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
                  total:
                    type: integer
                    description: Total number of comments
                  limit:
                    type: integer
                  offset:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Comments
      summary: Add a comment to a post
      description: |
        Create a new comment on a post. Requires authentication.
        Automatically creates notification for post author (unless commenting on own post).
        Updates post comment count.
      operationId: createPostComment
      security:
        - ClerkAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '201':
          description: Comment created successfully
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Rate limit (30 requests/minute for authenticated writes)
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                required:
                  - comment
                  - message
                properties:
                  comment:
                    $ref: '#/components/schemas/Comment'
                  message:
                    type: string
                    example: Comment added successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Post not found or deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Post not found
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/notifications/read-all:
    patch:
      tags:
        - Notifications
      summary: Mark all notifications as read
      description: |
        Mark all unread notifications as read for authenticated user.
      operationId: markAllNotificationsRead
      security:
        - ClerkAuth: []
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: All notifications marked as read
                  count:
                    type: integer
                    description: Number of notifications updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
security:
  - ClerkAuth: []